requirements.txt
numpy
pandas 
Flask
scikit-learn
seaborn
imblearn
python-box
PyYAML
-e .